% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_extra.R
\name{agent_kind}
\alias{agent_kind}
\title{Behavioral class of a rule}
\usage{
agent_kind(name)
}
\arguments{
\item{name}{Registry key.}
}
\value{
One of "maximizer", "meliorator", "heuristic".
}
\description{
Behavioral class of a rule
}
\examples{
agent_kind("melioration")
}
