% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{boltzmann_td_agent}
\alias{boltzmann_td_agent}
\title{Boltzmann Q-learning agent (softmax selection over Q)}
\usage{
boltzmann_td_agent(alpha = 0.1, gamma = 0.99, beta = 1, n_actions = 4L)
}
\arguments{
\item{alpha}{Learning rate.}

\item{gamma}{Discount factor.}

\item{beta}{Inverse temperature for selection.}

\item{n_actions}{Number of actions.}
}
\value{
A list of agent callables.
}
\description{
Boltzmann Q-learning agent (softmax selection over Q)
}
\examples{
boltzmann_td_agent()$select
}
