% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operant.R
\name{concurrent_schedule}
\alias{concurrent_schedule}
\title{Concurrent two-operandum schedule}
\usage{
concurrent_schedule(schedules = NULL, magnitudes = c(1, 1))
}
\arguments{
\item{schedules}{List of two schedule objects.}

\item{magnitudes}{Reinforcement magnitudes.}
}
\value{
An environment object with \code{reset} and \code{step}.
}
\description{
Concurrent two-operandum schedule
}
\examples{
concurrent_schedule(list(make_schedule("VI", 30), make_schedule("VI", 90)))
}
