% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurizer.R
\name{interval_featurizer}
\alias{interval_featurizer}
\title{Scalar (1-D) featurizer}
\usage{
interval_featurizer(n_bins = 20, lo = 0, hi = 1, dim = 1L)
}
\arguments{
\item{n_bins}{Number of bins.}

\item{lo, hi}{Range of the scalar.}

\item{dim}{Index of the scalar within the observation.}
}
\value{
A function mapping an observation to a character key.
}
\description{
Scalar (1-D) featurizer
}
\examples{
interval_featurizer()(0.5)
}
