% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featurizer.R
\name{lunar_low}
\alias{lunar_low}
\alias{lunar_high}
\title{LunarLander observation bounds}
\format{
Numeric vectors of length 8.
}
\usage{
lunar_low

lunar_high
}
\description{
LunarLander observation bounds
}
