% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_extra.R
\name{make_agent}
\alias{make_agent}
\title{Construct an agent by name}
\usage{
make_agent(name = "q_learning", n_actions = 2L, horizon = 20000L, ...)
}
\arguments{
\item{name}{Registry key.}

\item{n_actions}{Number of actions.}

\item{horizon}{Number of steps the agent will run (sets epsilon decay).}

\item{...}{Passed to the underlying constructor.}
}
\value{
A list of agent callables.
}
\description{
Exploration for the value-bootstrapping rules is scaled to \code{horizon}.
}
\examples{
names(make_agent("sarsa", n_actions = 2L, horizon = 5000L))
}
