% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operant.R
\name{make_schedule}
\alias{make_schedule}
\title{Build a schedule by name}
\usage{
make_schedule(kind = "VR", param = 5)
}
\arguments{
\item{kind}{One of "FR", "VR", "FI", "VI".}

\item{param}{Ratio or interval parameter.}
}
\value{
A schedule object.
}
\description{
Build a schedule by name
}
\examples{
make_schedule("VI", 30)
}
