% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{plot_cumulative_record}
\alias{plot_cumulative_record}
\title{Plot a cumulative record}
\usage{
plot_cumulative_record(cr, title = "Cumulative record")
}
\arguments{
\item{cr}{A cumulative record from \code{\link[=cumulative_record]{cumulative_record()}} or
\code{\link[=schedule_record_demo]{schedule_record_demo()}}.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Plot a cumulative record
}
