% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_differentiation_matrix}
\alias{plot_differentiation_matrix}
\title{Plot the differentiation matrix as a heatmap}
\usage{
plot_differentiation_matrix(dm)
}
\arguments{
\item{dm}{Output of \code{\link[=differentiation_matrix]{differentiation_matrix()}}.}
}
\value{
A ggplot object.
}
\description{
Plot the differentiation matrix as a heatmap
}
