% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_discounting}
\alias{plot_discounting}
\title{Plot a discounting fit}
\usage{
plot_discounting(delays, indiff, fit)
}
\arguments{
\item{delays}{Delays.}

\item{indiff}{Indifference amounts.}

\item{fit}{Output of \code{\link[=fit_discounting]{fit_discounting()}}.}
}
\value{
A ggplot object.
}
\description{
Plot a discounting fit
}
