% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatures.R
\name{plot_dra_fct}
\alias{plot_dra_fct}
\title{Plot DRA / FCT reallocation}
\usage{
plot_dra_fct(demo)
}
\arguments{
\item{demo}{Output of \code{\link[=dra_fct_demo]{dra_fct_demo()}}.}
}
\value{
A ggplot object.
}
\description{
Plot DRA / FCT reallocation
}
