% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_drl}
\alias{plot_drl}
\title{Plot DRL results}
\usage{
plot_drl(res)
}
\arguments{
\item{res}{Output of \code{\link[=drl_experiment]{drl_experiment()}}.}
}
\value{
A ggplot object.
}
\description{
Plot DRL results
}
