% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_extinction}
\alias{plot_extinction}
\title{Plot the extinction result}
\usage{
plot_extinction(df)
}
\arguments{
\item{df}{Output of \code{\link[=extinction_experiment]{extinction_experiment()}}.}
}
\value{
A ggplot.
}
\description{
Plot the extinction result
}
