% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_analysis.R
\name{plot_functional_analysis}
\alias{plot_functional_analysis}
\title{Plot a functional analysis}
\usage{
plot_functional_analysis(fa)
}
\arguments{
\item{fa}{Output of \code{\link[=functional_analysis]{functional_analysis()}}.}
}
\value{
A ggplot object.
}
\description{
Plot a functional analysis
}
