% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gym_toolkit.R
\name{plot_gym_dra}
\alias{plot_gym_dra}
\title{Plot a gym DRA reallocation}
\usage{
plot_gym_dra(demo)
}
\arguments{
\item{demo}{Output of \code{\link[=gym_dra]{gym_dra()}}.}
}
\value{
A ggplot object.
}
\description{
Plot a gym DRA reallocation
}
