% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gym_toolkit.R
\name{plot_gym_functional_analysis}
\alias{plot_gym_functional_analysis}
\title{Plot a gym functional analysis}
\usage{
plot_gym_functional_analysis(fa)
}
\arguments{
\item{fa}{Output of \code{\link[=gym_functional_analysis]{gym_functional_analysis()}}.}
}
\value{
A ggplot object.
}
\description{
Plot a gym functional analysis
}
