% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_herrnstein}
\alias{plot_herrnstein}
\title{Plot a Herrnstein hyperbola fit}
\usage{
plot_herrnstein(df, fit)
}
\arguments{
\item{df}{Output of \code{\link[=herrnstein_experiment]{herrnstein_experiment()}}.}

\item{fit}{Output of \code{\link[=fit_herrnstein_hyperbola]{fit_herrnstein_hyperbola()}}.}
}
\value{
A ggplot object.
}
\description{
Plot a Herrnstein hyperbola fit
}
