% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunarlander.R
\name{plot_lunar_convergence}
\alias{plot_lunar_convergence}
\title{Plot a LunarLander convergence demonstration}
\usage{
plot_lunar_convergence(demo)
}
\arguments{
\item{demo}{The list returned by \code{\link[=lunar_solved_convergence]{lunar_solved_convergence()}} or \code{\link[=lunar_best_policy_convergence]{lunar_best_policy_convergence()}}.}
}
\value{
A ggplot object.
}
\description{
Plot a LunarLander convergence demonstration
}
