% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_matching}
\alias{plot_matching}
\title{Plot the generalized matching law fit}
\usage{
plot_matching(fit)
}
\arguments{
\item{fit}{Output of \code{\link[=fit_generalized_matching]{fit_generalized_matching()}}.}
}
\value{
A ggplot.
}
\description{
Plot the generalized matching law fit
}
