% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_prob_matching}
\alias{plot_prob_matching}
\title{Plot probability-matching results}
\usage{
plot_prob_matching(res)
}
\arguments{
\item{res}{Output of \code{\link[=prob_matching_experiment]{prob_matching_experiment()}}.}
}
\value{
A ggplot object.
}
\description{
Plot probability-matching results
}
