% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_progressive_ratio}
\alias{plot_progressive_ratio}
\title{Plot progressive-ratio breakpoints}
\usage{
plot_progressive_ratio(df)
}
\arguments{
\item{df}{Output of \code{\link[=progressive_ratio_experiment]{progressive_ratio_experiment()}}.}
}
\value{
A ggplot object.
}
\description{
Plot progressive-ratio breakpoints
}
