% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_risk}
\alias{plot_risk}
\title{Plot risk-sensitivity results}
\usage{
plot_risk(res)
}
\arguments{
\item{res}{Output of \code{\link[=risk_experiment]{risk_experiment()}}.}
}
\value{
A ggplot object.
}
\description{
Plot risk-sensitivity results
}
