% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_extra.R
\name{plot_self_control}
\alias{plot_self_control}
\title{Plot self-control results}
\usage{
plot_self_control(res)
}
\arguments{
\item{res}{Output of \code{\link[=self_control_experiment]{self_control_experiment()}}.}
}
\value{
A ggplot object.
}
\description{
Plot self-control results
}
