% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_training}
\alias{plot_training}
\title{Plot LunarLander training returns}
\usage{
plot_training(res)
}
\arguments{
\item{res}{Output of \code{\link[=differentiate]{differentiate()}}.}
}
\value{
A ggplot.
}
\description{
Plot LunarLander training returns
}
