% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_trap}
\alias{plot_trap}
\title{Plot the melioration-trap result}
\usage{
plot_trap(exp)
}
\arguments{
\item{exp}{Output of \code{\link[=melioration_trap_experiment]{melioration_trap_experiment()}}.}
}
\value{
A ggplot.
}
\description{
Plot the melioration-trap result
}
