% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{regret}
\alias{regret}
\title{Regret against an optimal rate}
\usage{
regret(achieved, optimal)
}
\arguments{
\item{achieved}{Achieved reward rate.}

\item{optimal}{Optimal reward rate.}
}
\value{
The non-negative regret.
}
\description{
Regret against an optimal rate
}
\examples{
regret(0.4, 0.48)
}
