% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{run_app}
\alias{run_app}
\title{Launch the bundled Shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{Passed to \code{\link[shiny:runApp]{shiny::runApp()}}.}
}
\value{
Invisibly, the result of \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\description{
Launch the bundled Shiny app
}
\examples{
\dontrun{
run_app()
}
}
