% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_extra.R
\name{sarsa_agent}
\alias{sarsa_agent}
\title{SARSA agent (on-policy bootstrap)}
\usage{
sarsa_agent(
  alpha = 0.1,
  gamma = 0.99,
  eps_start = 1,
  eps_end = 0.05,
  eps_decay_steps = 150000,
  n_actions = 4L
)
}
\arguments{
\item{alpha}{Learning rate.}

\item{gamma}{Discount factor.}

\item{eps_start, eps_end, eps_decay_steps}{Epsilon schedule.}

\item{n_actions}{Number of actions.}
}
\value{
A list of agent callables.
}
\description{
Bootstraps with an action sampled from the current epsilon-greedy policy.
}
\examples{
names(sarsa_agent())
}
