% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\name{softmax}
\alias{softmax}
\title{Softmax with inverse temperature}
\usage{
softmax(x, beta = 1)
}
\arguments{
\item{x}{Numeric vector.}

\item{beta}{Inverse temperature.}
}
\value{
Probability vector.
}
\description{
Softmax with inverse temperature
}
\examples{
softmax(c(0, 1, -1, 0.5))
}
