% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{table_get}
\alias{table_get}
\title{Get a row from the table, defaulting to zeros}
\usage{
table_get(table, s, n_actions)
}
\arguments{
\item{table}{A table from \code{\link[=make_table]{make_table()}}.}

\item{s}{Character state key.}

\item{n_actions}{Number of actions.}
}
\value{
Numeric vector of length \code{n_actions}.
}
\description{
Get a row from the table, defaulting to zeros
}
