% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents_extra.R
\name{win_stay_lose_shift_agent}
\alias{win_stay_lose_shift_agent}
\title{Win-stay-lose-shift agent (operant heuristic)}
\usage{
win_stay_lose_shift_agent(n_actions = 2L, p_explore = 0)
}
\arguments{
\item{n_actions}{Number of actions.}

\item{p_explore}{Probability of a random action regardless of history.}
}
\value{
A list of agent callables.
}
\description{
Win-stay-lose-shift agent (operant heuristic)
}
\examples{
names(win_stay_lose_shift_agent())
}
